update: finialize designs - cleanup and fix
- finialize designs
- add flash loan to vault
- repkg
- add fee manager

+++ contracts/abstract/Multicall.sol
@@ -9,7 +9,7 @@ abstract contract Multicall {
     function multicall(bytes[] calldata data) public payable returns (bytes[] memory results) {
         results = new bytes[](data.length);
         
-        for (uint256 i; i < data.length;) {
+        for (uint i; i < data.length;) {
             (bool success, bytes memory result) = address(this).delegatecall(data[i]);
 
             if (!success) {
